+++ packages/syllst-xapi/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @syllst/xapi
 
+## 0.4.0
+
+### Minor Changes
+
+- Synchronized version with ecosystem-wide 0.4.0 release
+
+### Patch Changes
+
+- Updated dependencies
+  - @syllst/core@0.4.0
+
 ## 0.2.0
 
 ### Minor Changes